+++ okitweb/static/okit/views/compartment/js/resources/subnet.js
@@ -151,7 +151,7 @@ class SubnetView extends OkitContainerCompartmentArtefactView {
     }
 
     getLeftArtifacts() {
-        return [FileStorageSystem.getArtifactReference(), MountTarget.getArtifactReference()];
+        return [FileStorageSystem.getArtifactReference(), MountTarget.getArtifactReference(), InstanceConfiguration.getArtifactReference()];
     }
 
     /*